try hard to still use the clCreateCommandQueue() on 2.0
The OpenCL 2.0 standard replaces clCreateCommandQueue() with
clCreateCommandQueueWithParams(). However, some vendor backend libraries
stil do not provide it for 1.x platforms. This is what happened to my
nVidia GPUs on Debian. If I try to use it, it blows with SEGFAULT.

"Fortunately" the clCreateCommandQueue() is still there, it's just
marked as deprecated. This patch tries hard to still use the old
function as it was to this point.